refactor(st-09048): modularize multi-agent routing
